+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
  * Created `Invoke-WhsCIBuild` function for running a build as specified in a `whsbuild.yml` file.
  * Renamed `Pester` task to `Pester3`.
  * Fixed: `New-WhsAppPackage` puts packages in the root of the repository, not the `.output` directory.
+ * Added `WhsAppPackage` task to `whsbuild.yml` schema and `Invoke-WhsCIBuild`. This task creates application deployment packages.
 
 
 # 0.1.0